arquitectura en capas terminado
